// src/app/models/aspNetUsers.ts
export class AspNetUsers {
  id: string;
  userName: string;
  normalizedUserName: string;
  email: string;
  normalizedEmail: string;
  emailConfirmed: boolean;
  passwordHash: string;
  securityStamp: string;
  concurrencyStamp: string;
  phoneNumber: string;
  phoneNumberConfirmed: boolean;
  twoFactorEnabled: boolean;
  lockoutEnd: Date;
  lockoutEnabled: boolean;
  accessFailedCount: number;
  name: string;
  lastName: string;
  idGroup: number;
  masterDealer: boolean;
  inactive: boolean;
  idGroupName: string;
  password: string;
  customerId: number;
  rePasword: string;
}